Introduce concept of coder, extract ActiveRecordCoder
